commodity-weather-analysis/requirements.txt
# Core
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Time Series & Econometrics
statsmodels>=0.13.5
pmdarima>=2.0.3
arch>=5.3.1

# Machine Learning
scikit-learn>=1.1.0
xgboost>=1.7.0
tensorflow>=2.10.0

# Data
requests>=2.28.0
yfinance>=0.2.12
pandas-datareader>=0.10.0

# Visualization
plotly>=5.11.0
dash>=2.7.0
ipywidgets>=8.0.0
kaleido>=0.2.1

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
joblib>=1.2.0
python-dotenv>=0.21.0

# Jupyter
jupyterlab>=3.5.0
nbformat>=5.7.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
